# Conflicts:
#	data_manager.py

+++ main.py
@@ -19,27 +19,19 @@ def index():
     return render_template('index.html')
 
 
-@app.route("/api/add_new_board/")
-@json_response
-def insert_new_board():
-    return queries.insertNewBoard()
-
-
-@app.route("/api/boards/<int:board_id>/delete/")
-@json_response
-def delete_board(board_id):
-    return queries.delete_board(board_id)
-
-
-@app.route("/api/boards/<int:board_id>/<new_title>/")
-@json_response
-def update_board_title(new_title, board_id):
-    return queries.update_board_title(board_id, new_title)
-
-
-@app.route("/api/boards")
+@app.route("/api/boards", methods=['POST', 'GET', 'PUT', 'DELETE'])
 @json_response
 def get_boards():
+    if request.method == 'POST':
+        queries.insertNewBoard()
+    elif request.method == 'PUT':
+        board_id = request.json['id']
+        new_title = request.json['title']
+        queries.update_board_title(board_id, new_title)
+    elif request.method == 'DELETE':
+        board_id = request.json['id']
+        queries.delete_board(board_id)
+
     return queries.get_boards()
 
 
@@ -71,38 +63,6 @@ def get_statuses():
 @json_response
 def get_card_status(status_id):
     return queries.get_card_status(status_id)
-
-
-@app.route("/registration", methods=["POST", "GET"])
-def create_new_account():
-    if request.method == 'POST':
-        firstName = request.form['registrationFirstName']
-        secondName = request.form['registrationSecondName']
-        email = request.form['registrationEmail']
-        telephoneNumber = request.form['TelephoneNumber']
-        hashpassword = util.hash_password(request.form['registrationPassword'])
-        queries.create_account(firstName, secondName, email, telephoneNumber, hashpassword, )
-    return render_template('index.html')
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if request.method == "POST":
-        if request.form['user_name'] != '' and request.form['password_name'] != '':
-            login = request.form['user_name']
-            password = request.form['password_name']
-            print(queries.login(login)[0]['id'])
-            if util.verify_password(password, queries.login(login)[0]['password']):
-                session['name'] = queries.login(login)[0]['name']
-                session['email'] = queries.login(login)[0]['email']
-                session['id'] = queries.login(login)[0]['id']
-                return util.YOU_ARE_LOGGED_IN
-            else:
-                return util.INVALID_LOGIN_ATTEMPT
-        else:
-            return util.ENTER_ALL_VALUES
-    return render_template('index.html')
-
 
 
 def main():
